Fix: checking for `std::optional` availability must rely first on `__cpp_lib_optional` and in case it is not available fallback to checking headers ().
Previous implementation when used under C++17 was able to find `<experimental/optional>` header and use it as available optional which is useless.
If checking for `<optional>` first it causes errors under C++14 if such header is realy present (which is the case with compilers supporting both versions) so it is not the way out.